pyproject.toml
[project]
name = "webscraping-scrapy-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
    "scrapy>=2.12.0",
    "streamlit>=1.44.1",
]
